Update composer.json for packagist
The composer.json file was updated to include the project name ("scilone/symfonytoolboxbundle") and the license was changed from proprietary to MIT. The autoload configuration was also modified to properly map the "SymfonyToolboxBundle" namespace to the "src/" directory. These changes make the bundle easily identifiable and ensure it adheres to open source licence model for wider usage.