✨ added Monolog logging
This commit switches the application from using `App` to using `SymfonyToolboxBundle` to align with the project's standards. It also introduces `symfony/monolog-bundle` for logging capabilities. This logging package allows for more robust logging capabilities, which are critical for debugging and maintaining the application.

In addition, a `ProcessIdProcessor` is added to the Logger to provide unique process IDs for each logging event, which can help in tracing and debugging processes.