Remove UnknownSubscriptionException handling
The UnknownSubscriptionException class was deleted and references to it were removed in the SubscriptionFacade and AbstractPubSubMessagePullerCommand classes. This decision was based on the realization that throwing this exception was not necessary since Google Pubsub client library function 'subscription' automatically creates a new subscription when a non-existing one is requested. Therefore, the condition checking subscription's existence and throwing the UnknownSubscriptionException was redundant and eliminated to improve code simplicity and readability.